@benhoyt 2021-07-01
@inkel Side note: glad you like GoAWK. Out of interest, what are you using it for, or what project are you using it in? I always love hearing about how people are using it.