@inkel 2021-07-01
We're adding an AWK pipeline to Grafana Loki as part of an internal hackathon; you can check this PR: https://github.com/grafana/loki/pull/3906

So far it's been working really great! I don't know if it will be added to Loki or not (you can only execute the program for one line at a time, so no `BEGIN` nor `END`).

I might have found another improvement but I need to work out the details yet. I'll probably let you know tomorrow about it, either via an issue or PR of I find the time to code it 😬