@benhoyt 2021-07-01
Interesting, thanks!

I took a closer look at your changes, and there are a couple of issues:

* A [`rand.Source`](https://golang.org/pkg/math/rand/#NewSource) is not sharable between multiple goroutines, so I don't think we can reuse a global source for all calls to `ExecProgram`. Users should be able to use ExecProgram safely from multiple goroutines. This also changes the behavior (apart from the fact it's not concurrency-safe): two concurrent interpreters will be using the same stream of random numbers, so each program will see a different stream than if you were running a single program.
* `awk` and `gawk` (and the current `goawk`) both have this behavior where the seed is the same each time. (This is documented with a "caution" in the Gawk docs, see `rand` [here](https://www.gnu.org/software/gawk/manual/html_node/Numeric-Functions.html).) That's why I used the static seed approach originally. You can use `srand()` in your AWK program if you want a different stream each time.

Note that `ExecProgram`, as you've seen, has a pretty high startup cost. It does a whole bunch of allocations and setup, and isn't intended to be called line by line. Is there a way you could call `ExecProgram` once and pass your logs as input to the `interp.Config.Stdin` reader line by line? (You might have to use concurrency or a goroutine to do the sending.) I think this would be a lot more efficient, as it would avoid the `ExecProgram` startup cost each for each log line.

If we did want to add the ability to make the seed different each time (and I'm not opposed to that as a configuration option), I suggest we add a new field to `interp.Config`, something like `Randomize bool`, which if set to true, uses a time-based input as the seed in the `ExecProgram` setup.