@inkel 2021-07-06
Hey, thanks for replying so thoroughly! I appreciate that.

I totally understand your points, and in fact I was surprised to learn that indeed a static seed is the standard for all AWK versions.

I'll look into your suggestions, although if implemented then I'll probably do it in another PR. I'm on my phone right now so it's hard to think about the code 😺 

As for the startup cost of `ExecProgram` profiling that was what lead me to this PR, so I'm pretty aware of it. Sadly the way the app where we are implementing GoAWK works this way of having to call that method once per line, although as I'm writing this an idea is coming to mind, so I might get around it. I'll keep you posted!

Thanks again for your time and commitment with this library!